Remove paasta generate-pipeline command
It's probably time to rip the bandaid off completely now. At this point,
the only services that could be using this are probably:
    * using customized pipelines (which this doesn't support)
    * or are doing something non-standard (and probably not changing at all)
    * {insert another reason here}

In any case, it's probably time to remove this escape hatch and, if
necessary, make our current standard pipelines support any of the
esoteric usages of generate-pipline or move jobs to custom Jenkinsfiles.