Prevent scaling down services if the utilization > setpoint
The affects the moving_average forcast policy.  If the usage was low for
several minutes, this could affect the average enough to prompt scaling
down in more than one service autoscaler run.
This prevents that case by refusing to scale down if the current usage
is greater than the setpoint (or the high point of the good_enough_window)