Set disk quota for k8s
Set `ephemeral-storage` resource and limit for pods according to the
configuration, setting the disk quota as a result.

Please note that that is a soft disk quota which doesn't prevent a pod
from using more disk space than requested.  Instead of that, Kubernetes
evicts the violating pod, which then either rescheduled to another node
or becomes `Evicted`.  Please see the corresponding KEP for more
information:
https://github.com/kubernetes/enhancements/blob/master/keps/sig-node/20180906-quotas-for-ephemeral-storage.md