Add kubernetes status for cassandraclusters
This adds the info on the statefulset and pods that are added to
KubernetesDeploymentConfig status API. These are also useful for
cassandra since the operator creates a stateful set of a similar type.
To make this work I had to add namespace support in more places and I
set the namespace in kubernetes_tools to paasta and in
cassandracluster_tools to paasta-cassandraclusters.