Upgrade marathon-python to a version that returns naive datetimes.
This fixes a bug in the Paasta API where the POSIX timestamps relating
to Marathon apps and tasks were off by the difference between local time
and UTC (but in the opposite direction). Because the API now produces
accurate timestamps, the CLI no longer needs to adjust the timestamps
back in the other direction.

This commit also cleans up code from several places that is meant to
deal with the naive datetime objects previously returned by
marathon-python.