check_flink_services_health sends a single event per service instance
FLINK-1982

The event is going to be a combination of all the 4 checks we have in
place.

In order to re-use as much as possible the code in monitoring_tools.py,
I added an helper function where I factored out the logic for the
under-replication check without the actual sending of the event. I am
happy to change the tests for monitoring_tools.py too if you want me to;
I didn't do it to minimize the overall lines of code changed given that
I only factored out code and added a return statement.