Add second docker wrapper
This is kinda ugly but it does mean that docker inspects will be fast
again. With this 100 docker inspects ~ 15s without ~ 105s

There might be a better fix but just putting this up so we can consider
it.

The motivation is that mesos times out on slow docker inspects and the
python wrapper can be very slow at importing all the python code.