Don't try to retrieve authors from GitHub repos with gitolite commands
The old way of finding authors is to call a gitolite command that
finds the authors between two commits. For services on GitHub, we pass
in the authors when making a mark-for-deployment call, so there is no
benefit to trying to call this gitolite command on a GitHub repository.

This change was prompted by the noisiness of timed deploys for shas
already on master not having authors passed in, and complaining about
not being able to find any authors from GitHub as a fallback