Refactor instance listing into a single function
We can simplify the tron logic because we don't have to worry about
jobs and actions being lists. They are always dicts now.

This is in preparation for changing the way we read instance configs
(PAASTA-16300). Ideally that would be a single function, e.g.
read_instance_configs, that all other tools can use.