Fixes to the kube_state_metrics_collector's config schema
Previously to join the labels from one metric to another, we specified a mapping
of the label metric name to 2-element list of the label metric's join key to the
join key to look for on other metrics.  This does not work if we want to join
the same label metric to multuple join keys since the label metric would have to
be listed twice in the mapping, which causes Python to overwrite one of them in
the resulting dict.  The new schema allows for this possibility by separating
the source key from the label metric from a list of potentially multiple
destination join keys.