Increase default cpu/mem for tasks
With automatic rightsizing, the goal is for most services to not
need to set cpu/mem in their configs. When a service first starts,
we won't have any usage data to make recommendations on, so
the defaults should be generous. The rightsizer will scale them
down as appropriate after that. This maintains the 1:4 ratio
of cpus to mem, which matches the makeup of our clusters pretty
well.