Support envoy in a readiness check
`is_pod_healthy_in_proxy.py` and `check_proxy_up.sh` are new readiness
checks that can healthcheck at both smartstack and envoy backends.

The smartstack healthcheck is exactly the same as the current
`is_pod_healthy_in_smartstack.py` used by `check_smarstack_up.sh`.

The envoy healthcheck checks for envoy backends.

Toggles:
* First, the service must be enabled in `envoy_toggles.yaml`
* Second, the service can also be enabled for full mesh in
  `full_mesh_envoy_toggles.yaml`

If a service is enabled for envoy, then the healthcheck examines envoy
for a host in the `egress_cluster` that corresponds to this service and
have matching IPs for the pod IP.

If a service is enabled for full_mesh envoy, then the healthcheck
additionally examines envoy for a single healthy host in the `ingress_cluster`.
Furthermore, the hosts in the `egress_cluster` must have IPs matching
the paasta host fo the service (instead of pod IP).

The envoy healtcheck is additionally gated by the
`bounce_healtcheck_params` `check_envoy` config toggle. By default, this
is False.

This readinesss will not be used until:
* sidecar containers support paasta-tools by mounting paasta related
  paths with volume mounts
* sidecar container image is updated with paasta-tools containing this
* commit
* the readiness probe for sidecar containers is refactored to supports
  calling the readiness script appropriately