Fix error where setup.py was trying to install paasta_tools/contrib/__pycache__ as a script.
The import from a contrib script in firewall.py causes python to create a pyc file under paasta_tools/contrib/__pycache__.
On the next run of setup.py, this __pycache__ directory would be passed in the scripts parameter to setup().
setup() would then fail to install a directory as a script with:

    error: [Errno 21] Is a directory: '/nail/home/krall/pg/paasta/paasta_tools/contrib/__pycache__'

As it turns out, this import shouldn't be here anyway -- it was removed in e9fdebcf2, but in a different branch it was moved in eae9a97a2.
The merge commit, 78e57cf35, ended up with the import still existing in the end.