Implement Kubernetes versions of the acceptance tests using kind
The pre-existing acceptance testing was running against a Dockerized
Mesos cluster, which is going to start causing problems as we want to
test Kubernetes-specific functionality.  This change adds in a
Kubernetes cluster via `kind`, and runs the autoscaler acceptence tests
against it.

We also fix a small issue where the metrics collector crashes if it
can't load a PoolManager for a particular pool; now it will just log the
error and continue.