Small-ish change to stick a little glue in place for the PendingPods signal
The pending pods signal will now optionally just return a list of the
unscheduled pods in the cluster (protip: don't do this yet or
literally nothing will autoscale at all).

If you pass in the `per_pod_resource_requests` parameter to the signal,
it will just find the unschedulable pods and return them in a list.
This list should get passed in to the
`_compute_new_resource_group_actions` function from
https://github.com/Yelp/clusterman/pull/51.

That function will then call the
`PoolManager._filter_scale_up_options_for_pod` in order to determine
what nodes options the pod could run on (note that, by passing in a
`ClusterNodeMetadata` object with `allocated_resources` set we can
determine if the pod will fit on a pre-existing node).