Remove all dependence on DynamoDB metrics from the internal signal
The metrics collector batch is incredibly slow and has been causing some
scaling woes, so this removes our dependence on the metrics emitted by
that batch, and just queries the Kubernetes apiserver directly for all
metrics.

A consequence of this is that the signal no longer looks for the `boost`
metrics, but this seems fine because boost hasn't worked on Kubernetes
for a long time anyways.  At some point in the future we can either
deprecate boost in favor of `clusterman manage` or re-add the boost into
this code in a better way.