make some changes to try to reduce memory usage
The main change here is that we don't require the autoscalers to track
all the pods, just the ones that belong to the current pool.  This
should hopefully reduce the amount of memory used by _most_ of the
autoscaler instances.

To accomplish this, we just filter all the pods that are a) running on a
node in our pool, b) has a node selector for nodes in our pool, or c)
has an affinity for nodes in our pool.

We still cache everything for the metrics collector, so these will still
need large memory utilization.