This change fixes two bugs found in the autoscaling behaviour
1) The first, and less serious, is that we weren't reloading the pool
state until _after_ we'd already evaluated our signals, which means that
for the internal version of the signal (which doesn't rely on the
metrics store), we were using the data from the _last_ autoscaling run
instead of the current one.

2) The second is a bit larger: it turns out we were never actually
scaling based on pending pods, because pending pods were always getting
marked as "Unknown".  The reason for this is that we were comparing the
pending pods request to the total allocatable capacity on the node,
instead of what's remaining, so Clusterman never realized that pods
potentially couldn't fit on a given node.

This change fixes both of these behaviours.  Since these both change how
the scaling behaviour actually happens, we'll want to roll it out
slowly.