Add support for k8s secret sync to namespaces
To support paasta secrets in operator workloads we need to sync secrets
for services that run in other namespaces into those namespaces. This
adds mapping for workloads to k8s namespaces and changes the sync script
to list all services and find any instance types and syncing to those
instance types namespaces.