Make get_kubernetes_metadata use actual node pool
This fixes an assumption that pods use just a node selector. Some of the
cassandra Pods use node affinity and in practice a lot of things can
impact the node/pool that a Pod actually ends up on. This checks the Pod
to see what node it is scheduled on and gets the pool from the Nodes
labels (which I think is more reliable).

Note I tested this on kubestage but I don't know the impact if this
breaks something so please warn me if I need to be careful.