Make pods with a prometheus_port routable
Prometheus shards (currently deployed outside of PaaSTA) that need to
scrape metrics from a service need to be able to contact that service.

The only way to do so currently is to add a port in Smartstack in order
to make the pod ip routable (via the `paasta.yelp.com/routable_ip`
label). That port is unused and just serves to enable this behaviour.

While reserving unused ports in Smartstack is not too big a deal,
Smartstack/PaaSTA will launch a hacheck sidecar container which will run
extra healthchecks and use extra resources.
For Kubernetes operator pods, only one pod per deployment (the leader)
will have a Prometheus endpoint running (and able to accept
healthchecks). The other pods are therefore incorrectly considered
unhealthy, while they are just waiting for the lock to become leader.

This commit makes pods with a `prometheus_port` have their ip routable,
to remove the need for an unused Smartstack port and healthchecks.
Prometheus shards are able to gracefully deal with pods that do not
return a response (e.g: non-leader pods in an operator).