Remove GitHub changelog + make tagging/pushing safer
We don't normally release things with a GitHub token so the changelog is
several years out of date.

Additionally, force-tagging isn't particularly safe and neither is
pushing all the tags that exist locally. Instead, tell users to tag
normally and then give them the command-line to atomically push the
version bump and tag.