[local-run] Conditionally set MESOS env vars + unmangle docker hostname
In PAASTA-17466, an internal user reported that PAASTA_HOST is not a
resolvable hostname. This is because we're unconditionally mangling that
value and adding a uuid that dates back to when PaaSTA ran on Mesos.

This PR stops unconditionally setting certain Mesos-only environment
variables and additionally stops mangling the hostname we pass to Docker
if Mesos environment variables aren't set