Cleanup failed Pods in prune_completed_pods.py
The current version of this script doesn't capture Pods that end up in
Error or CreateContainerError states, and thus these end up sticking
around for a while.

There's no direct way to get what Pods are in this state, but we can use
the Failed phase as a proxy.

For now, the new arg I've added is optional until we've rolled this
version out everywhere and have updated all callsites to pass in
--error-minutes.

I'm planning on setting --error-minutes to a somewhat high value in case
we want to debug this using kubectl (so probably a couple hours?)