Make paasta_prune_completed_pods less likely to crash
We've been seeing two weird issues in one of our stage clusters:
1) Pods in an Error state with no PodScheduled condition
2) Pods returned by list_namespaced_pod 404ing when we attempt to delete
   them through delete_namespaced_pod

I was unable to repro 1) even after running this script in dry-run mode
for several days and was similarly unable to repro 2).

Since this is just a cleanup script, I decided to just log that we hit
an exception in these cases and continue on under the assumption that
these will either be cleaned up once the k8s API stops being flaky or
we'll be able to actually get more information on issue 1) with the
added logging and fix the issue at the root.