Skip clusterman metrics for spark-on-k8s
This is causing us to scale up our legacy mesos pools when k8s jobs
launch.

This does also mean that we won't print any cost warnings, but we should
probably get Core ML to refactor this further once we've stopped the
bleeding with this PR to get that feature back.