Status: write all available status objects
The Cassandra status object contains data for both kubernetes and
cassandracluster. This change will invoke the status writer for all
available status metadata. In the case of Cassandra clusters, it would
print the output of both kubernetes and cassandracluster writers.

The motivation is that the Kubernetes writer provides information that
are not available in the custom Cassandra status writer. So, we ideally
want to see the output of both of them.

Context: PAASTA-17617