Allow using preferred node affinity on its own
There was a subtle bug in the original implementation where setting only
a preferred node affinity would result in a required node affinity being
created with an empty match_expressions - which in our infra does not
match any nodes.

This patch ensures that we only created a required node affinity if
there is actually one configured.