Make get_running_task_allocation include all namespaces by default
For the most part, we're not really excluding many namespaces (most of
our namespaces start with paasta).

That said, now that we allow folks to use any arbitrary namespace, we
can run into issues where a namespace is not in our allowlist and then
autotune silently stops working.

I don't forsee this adding too much additional data (and if it does, we
can filter out noisy namespaces).