Scale with ready pods count rather current replicas for uwsgi
We had some surprising behavior recently where an autoscaled app was
crashlooping and the HPA did not keep the desired scale as-is - we
believe that this is due to a mismatch in how we calculate the value we
send to k8s:

The existing PromQL divides by current replica to turn things into a
utilization percentage such that when the HPA multiplies later on by the
number of current replicas, we reach our target replicas.

However, the HPA will actually multiply by the number of ready replicas,
leading to a mismatch.

(h/t to @krall for wording things much better than I could)