Speed up the default preStop hook to allow for graceful shutdown
The default PaaSTA preStop hook sleeps for 30s - which is exactly the
same as the default termination grace period, therefore preventing
anything that gets this default hook from being able to gracefully
shutdown.

We could (should?) look into simply not adding a hook for things that
don't need one (i.e., not mesh-registed and no custom hooks), but in the
meantime adding a silly message doesn't seem particularly harmful.