Skip autotune merging for deleted instances
It's possible for autotune results to include instances that no longer
exist in a given file: either because the instance has been completely
deleted or moved to another cluster.

In these cases, it should be safe to simply skip the config merge since
there's nothing to merge with.

We could technically still do the config update/merge here and simply
skip the resource clamping - but imo, we should simply bail out early.