tron: update schema for bulkdata + correctly pass extra volumes
We collectively missed that get_volumes() gets called on the
TronActionConfig and not the TronJobConfig. Furthermore, we missed that
we weren't actually calling get_volumes() for normal Tron jobs -
instead, we were only calling get_extra_volumes() and relying on the
Tron `default_volumes` config to add the usual default volumes (which
differs from how services work: we always call get_volumes() as we don't
distinguish between user-provided volumes and default ones when
generating the config there).

I've tested this on tron-infrastage (as well as ensured tests pass for
this commit), so this should be safe.