Update regular filename reference in autotune files
We've generally decided that we'll be leaving files as eks-*.yaml - and
while there's some additional steps that need to be taken, those will
need to happen once we've gotten rid of all of the kubernetes-*.yaml
files.

As we still have some unmigrated operators, that means we still need to
wait. However, we can go ahead and update the comment in the autotune
files since those have recently caused some confusion (and assume that
anyone looking at autotune files is looking at them for non-operator
services)