Add service and instance annotations to tron pods
We can currently only figure out what service/instance a log belongs to
by looking at the k8s labels for the emitting pod, but label values are
quite limited in length and we've got some pretty large job and/or
action names, which means that the instance label for a large chunk of
tronjobs ends up getting truncated.

Solution: annotations! these have a signficantly higher limit (256kb)
and they can still be read by our otel collector - the only downside is
that annotations cannot be used for filtering, but that's fine :)