Implement v2 of prometheus-based uWSGI autoscaling.
This uses the AverageValue style of target for the HPA rather than a target Value,
which means that the HPA will divide by the number of currently-running pods for
us, instead of us having to do that division. This should lead to more accurate
scaling. (Internally, the HPA then takes the metric and multiplies it by the number
of pods later, to get the desired pod count. We've seen times when the number we
divide by is different than the number the HPA multiplies by. By having the HPA
handle the division, hopefully it divides by the same number it multiplies by later.)

Since we're making changes, we may as well implement the setpoint as the target
averageValue, instead of having the HPA target be 1.0 and having our prometheus
query divide by the setpoint (as v1 uwsgi autoscaling does currently. We implemented
it this way because we used to have `offset`, which did hard-to-describe things to
the autoscaling behavior, which made it impractical to use setpoint as the HPA's
target.)