the remote-run missing bits (#4028)
This was meant to be included in the last release but I (luisp) don't know how to read and merged things in the "wrong" order. 

This correctly merges in the following PRs:

* cleanup for remote-run resources (#4024)

Clean up logic for the (meant to be) ephemeral resources which are created in remote-run invocations (#4022).

* new remote-run cli (#4025)

Last portion of #3986, with the updated logic to reflect the updated API conventions introduced in #4022.