Use 22.04 GHA runners (+ always use py3.8) (#4047)
GitHub is deprecating the 20.04 runners, and for some reason we're setting up py3.7 even though we're expecting 3.8.

Note: this required some packages bump: for some reason, `check-requirements` was reporting
that not all packages were installed -- even though the preceding tox loglines reported that
the "missing" packages were installed.

I tried to be pretty conservative with the bumps since we need to do a mass-upgrade at some
point in the future and it wasn't worth dealing with breaking changes and whatnot at the 
moment