Document lack of boto_keys in local-run (#4107)
I initially just added this in the local-run --help output, but then I went
ahead and also updated the rST files.

As part of this, I noticed that iam_role wasn't documented, and added
an entry for that as well.

There's probably some additional keys that we should add here, but these
are the first two that came to mind.

I've also gone ahead and replaced the kubernetes-[clustername].yaml
heading with eks-