Validate smartstack.yaml in `paasta validate`
We currently do this in an internal repo (`py-gitolite`), but I don't
see why we don't just do this in `paasta validate` - especially since
moving this check left means that folks are alerted to broken config
earlier rather than when they attempt to integrate.

The smartstack_schema.json in this repo was behind the one in
py-gitolite, so I've gone ahead and made them match.

The majority of the added code (outside the tests, which I had Claude
Code regenerate for me) is more-or-less taken straight from py-gitolite
with minimal changes.

I ran `paasta validate` over all soaconfigs and verified that there are
no new failures when run with this code.