add support for host aliases in paasta
in #dar-2776 we needed to hardcode dns resolution to an aws service
because of an upstream outage. this allows use of k8s HostAliases, which
add individual entries to /etc/hosts inside the pod.
https://kubernetes.io/docs/tasks/network/customize-hosts-file-for-pods/

we also add a ticket_ref key to the config to ensure we've got an audit
trail - this should only really be used in emergencies when we need to
do something temporarily, or when we need deployment-level granularity
on the change. we'd normally do this at the host-level dns resolver.